backend/requirements.txt
Package                       Version
----------------------------- --------
asgiref                       3.4.1
Django                        3.2.8
django-cors-headers           3.10.0
djangorestframework           3.12.4
djangorestframework-simplejwt 5.0.0
Pillow                        8.4.0
pip                           21.3
PyJWT                         2.3.0
pytz                          2021.3
setuptools                    57.0.0
sqlparse                      0.4.2
typing-extensions             3.10.0.2
wheel                         0.36.2
